requirements.txt
allure-python-commons

playwright

pytest

pytest-playwright

allure-pytest